US70515; segregating export functionality
